requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.7
langgraph>=0.0.20
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
requests>=2.31.0
pydantic>=2.0.0
aiohttp>=3.9.0
